Kona: Rename fstab.qti to fstab.qcom
qcom product string comes from ABL.
And it is populated in ro.hardware and hence in
ro.boot.hardware.

At multiple places, there is getproperty for
ro.hardware and ro.boot.hardware to get qcom string.

There are so many filenames with qcom string in it.

In short, it is grandfathered requirement to keep it as
fstab.com.

Change-Id: I2bc44c450db596adb51941146b49126d8f01c5cc